requirements.txt
streamlit
arxiv
sentence-transformers
faiss-cpu  
transformers
torch
python-dotenv
tiktoken 
requests
